feat(cadastro-usuario): adicionar cadastro de usuario.

+++ rede-social/rede-social-service/src/main/java/br/com/crescer/repository/PerfilRepository.java
@@ -1,12 +1,13 @@
 package br.com.crescer.repository;
 
 import br.com.crescer.entity.Perfil;
+import java.math.BigDecimal;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 
 /**
  * @author vinicius.ambrosi
  */
-public interface PerfilRepository extends PagingAndSortingRepository<Perfil, Long>{
+public interface PerfilRepository extends PagingAndSortingRepository<Perfil, BigDecimal>{
     Perfil findOneByDsEmail(String email);
 }
